# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ZYRO API
  summary: Multi-tenant e-commerce platform with built-in AI content generation and shopping assistant.
  description: |
    This is the API contract draft for Phase 0, Module 3 of the ZYRO (ShopMind AI) implementation
    plan. It consolidates the endpoints sketched per-phase in `documentation/Implementation_Plan.md`
    into one consistent, code-gen-friendly contract so frontend and backend work (owned by different
    students per the scope document's Section 12 module split) can proceed in parallel without
    integration surprises.

    ## Multi-tenancy

    Every tenant-scoped resource is nested under `/stores/{storeId}/...`. `storeId` must always be
    supplied explicitly — this mirrors the tenant-scoping design from Phase 0 Module 1 (Prisma
    middleware and the Mongoose `tenantScopePlugin` both refuse a query with no tenant filter), so the
    API surface makes the same requirement visible to callers rather than inferring it silently.

    ## Auth model

    - `bearerAuth` — JWT access token. Required for all merchant/staff actions and any action tied to
      a logged-in customer account.
    - `guestSession` — an opaque guest session id (issued on first cart interaction, carried in the
      `X-Guest-Session-Id` header). Storefront browsing, cart, and checkout endpoints accept either
      `bearerAuth` OR `guestSession`, since Section 6 of the scope document requires guest checkout to
      work without an account.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /stores/{storeId}:
    parameters:
      - $ref: "#/components/parameters/StoreId"
    get:
      tags: [stores]
      operationId: stores_get
      summary: Get a store's public profile
      description: Public — used by the storefront to render branding before any auth check.
      security: []
      responses:
        "200":
          description: Store profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Store" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /stores/{storeId}/staff:
    parameters:
      - $ref: "#/components/parameters/StoreId"
    get:
      tags: [stores]
      operationId: store_staff_list
      summary: List staff accounts for a store
      description: Owner-only.
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StaffMember" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [stores]
      operationId: store_staff_create
      summary: Invite a staff account with scoped permissions
      description: Owner-only. Permissions gate access to specific routes (Section 6/7.2 role-based access).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, permissions]
              properties:
                email: { type: string, format: email }
                permissions:
                  type: array
                  items: { $ref: "#/components/schemas/StaffPermission" }
            examples:
              catalog_staff:
                value:
                  email: jordan@auroragoods.example
                  permissions: [products_write]
      responses:
        "201":
          description: Staff account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffMember" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /stores/{storeId}/products:
    parameters:
      - $ref: "#/components/parameters/StoreId"
    get:
      tags: [catalog]
      operationId: products_list
      summary: Browse or search a store's product catalog
      description: |
        Public. Supports plain browsing (category filter, pagination) and keyword search via
        the `q` parameter against the MongoDB text index on title/description
        (Implementation_Plan.md Phase 3) — folded into this collection endpoint rather than a
        separate `/search` path, since a search is just a filtered list, not a distinct resource.
      security: []
      parameters:
        - name: q
          in: query
          description: Keyword search across product title and description.
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Product page
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
                  pagination: { $ref: "#/components/schemas/PaginationInfo" }
    post:
      tags: [catalog]
      operationId: products_create
      summary: Create a product
      description: Requires the `products_write` staff permission (or owner).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /stores/{storeId}/uploads/images:
    parameters:
      - $ref: "#/components/parameters/StoreId"
    post:
      tags: [uploads]
      operationId: uploads_image_create
      summary: Upload a single image and get back its URL
      description: |
        `ProductInput.images` and `store_branding_update`'s `logoUrl` both take
        already-hosted URLs, not raw file data — this endpoint is what produces one.
        Added during Phase 1 Module 4 implementation: the original contract modeled
        product/branding images as pre-existing URLs without ever specifying how a URL
        comes to exist, which is a real gap once you try to build the upload flow.
        Requires the `products_write` staff permission (or owner). Stores the file to
        local disk for this prototype (Implementation_Plan.md Phase 1 names this as the
        deferred choice between S3-compatible object storage and local disk) and returns
        its served URL — never the binary itself, which is never persisted to a database.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: JPEG, PNG, or WebP, up to 5 MB.
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url: { type: string, format: uri }
        "400":
          description: Missing file, or an unsupported type/size
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ── REVIEWS ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    guestSession:
      type: apiKey
      in: header
      name: X-Guest-Session-Id
  parameters:
    StoreId:
      name: storeId
      in: path
      required: true
      description: Tenant identifier. Every operation on a tenant-scoped resource must supply this explicitly.
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Caller lacks the required staff permission
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      description: RFC 7807 problem details.
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: Stable identifier for this error category.
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
      examples:
        quota_exhausted:
          value:
            type: https://zyro.dev/errors/ai-quota-exhausted
            title: Monthly AI usage quota exhausted
            status: 402
            detail: This store has used all 100 AI generations included this month.
    PaginationInfo:
      type: object
      required: [total, limit, offset]
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...
    Store:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        logoUrl: { type: string, format: uri }
        themeColor: { type: string }
    StaffPermission:
      type: string
      enum: [products_write, orders_write, discounts_write, analytics_read]
    StaffMember:
      type: object
      required: [id, userId, permissions]
      properties:
        id: { type: string }
        userId: { type: string }
        email: { type: string, format: email }
        permissions:
          type: array
          items: { $ref: "#/components/schemas/StaffPermission" }
    ProductInput:
      type: object
      required: [title, price, stock, category]
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 5000 }
        price: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
        category: { type: string }
        images:
          type: array
          items: { type: string, format: uri }
          maxItems: 10
    Product:
      allOf:
        - $ref: "#/components/schemas/ProductInput"
        - type: object
          required: [id, storeId]
          properties:
            id: { type: string }
            storeId: { type: string }
            aiDescriptionStatus:
              type: [string, "null"]
              enum: [draft, published, null]
      examples:
        earbuds:
          value:
            id: prod_01HXYZ
            storeId: store_01ABCD
            title: Wireless Earbuds Pro
            description: Crisp sound, all-day battery, and a snug fit for workouts or commutes.
            price: 49.00
            stock: 42
            category: Electronics
            images: ["https://cdn.zyro.dev/products/earbuds-1.jpg"]
            aiDescriptionStatus: published
...